+++ migrations/sqlite/20260323000000_add_html_data_repo_name.down.sql
@@ -1,0 +1,4 @@
+-- Add down migration script here
+PRAGMA foreign_keys = OFF;
+
+PRAGMA foreign_keys = ON;