+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to a _modified_ form of _[Semantic Versioning][semver]_
 - Add redirects for dynamic routes ([#97])
 - Added _date endpoint from portal into stele ([#100]) 
 - Added _publication endpoint for historical version ([#104])
+- Added _date endpoint from portal into stele ([#100])
+- Add `--force` flag and automatic inconsistency detection to `stelae update` ([#108])
 
 ### Changed
 
@@ -25,6 +27,28 @@ and this project adheres to a _modified_ form of _[Semantic Versioning][semver]_
 [#100]: https://github.com/openlawlibrary/stelae/pull/100
 [#98]: https://github.com/openlawlibrary/stelae/pull/98
 [#97] https://github.com/openlawlibrary/stelae/pull/97
+
+## [v0.6.6]
+
+### Fixed
+
+- Fix commit insertion for multiple data repos of the same type ([#107])
+
+[#107]: https://github.com/openlawlibrary/stelae/pull/107
+
+## [v0.6.5]
+
+### Added
+
+- Track html data repo per publication for multi-repo support ([#106]) 
+
+### Changed
+
+### Fixed
+
+### Removed
+
+[#106]: https://github.com/openlawlibrary/stelae/pull/106
 
 ## [v0.6.4]
 
@@ -304,7 +328,9 @@ and this project adheres to a _modified_ form of _[Semantic Versioning][semver]_
 
 ### Removed
 
-[Unreleased]: https://github.com/openlawlibrary/stelae/compare/v0.6.4...HEAD
+[Unreleased]: https://github.com/openlawlibrary/stelae/compare/v0.6.6...HEAD
+[v0.6.6]: https://github.com/openlawlibrary/stelae/compare/v0.6.5...v0.6.6
+[v0.6.5]: https://github.com/openlawlibrary/stelae/compare/v0.6.4...v0.6.5
 [v0.6.4]: https://github.com/openlawlibrary/stelae/compare/v0.6.3...v0.6.4
 [v0.6.3]: https://github.com/openlawlibrary/stelae/compare/v0.6.2...v0.6.3
 [v0.6.2]: https://github.com/openlawlibrary/stelae/compare/v0.6.1...v0.6.2